(fix) Made the IndexQueryset respect a custom index name (#5)
The IndexQueryset was not respecting the custom index name
set in the Meta class of the model.
This was causing the queryset to always use the
default index name.

This commit fixes this issue by making the IndexQueryset
respect the custom index name set in the Meta class of the model.