fix: Bundled Meilisearch Requests (#7)
This commit bundles the Meilisearch requests to
reduce the number of requests made to the Meilisearch server.
This is done by utilizing the `update_settings` method on the
client SDK to update the settings in a single request. However,
index creation is still a separate request.

This also handles the fix for checking if the index exists
as the current implementation would always return false.